+++ app/src/main/report/actions/reportPageActions.ts
@@ -1,6 +1,7 @@
 import { Dispatch } from "redux";
 
 import { reportsActions } from "./reportsActions";
+import {userActions} from "./userActions";
 
 export const reportPageActions = {
 
@@ -9,6 +10,7 @@ export const reportPageActions = {
             dispatch(reportsActions.setCurrentReport(props.report));
             dispatch(reportsActions.getReportVersions(props.report));
             dispatch(reportsActions.getVersionDetails(props.report, props.version));
+            dispatch(userActions.getReportReaders(props.report));
         }
     }
