+++ app/src/main/report/components/Sidebar/PublishSwitch.tsx
