+++ app/src/main/report/reducers/reportAppReducers.ts
@@ -1,19 +1,22 @@
-import { combineReducers } from "redux";
-import { reducer as formReducer, FormReducer } from "redux-form";
+import {combineReducers} from "redux";
+import {FormReducer, reducer as formReducer} from "redux-form";
 
-import { authReducer, AuthState } from "../../shared/reducers/authReducer";
-import { reportsReducer, ReportsState } from "./reportsReducer";
+import {authReducer, AuthState} from "../../shared/reducers/authReducer";
+import {reportsReducer, ReportsState} from "./reportsReducer";
+import {usersReducer, UsersState} from "./userReducer";
 
 export interface ReportAppState {
     auth: AuthState;
     form: FormReducer;
     reports: ReportsState;
+    users: UsersState;
 }
 
 const reducers = combineReducers({
     auth: authReducer,
     form: formReducer,
-    reports: reportsReducer
+    reports: reportsReducer,
+    users: usersReducer
 });
 
 export default reducers;
